requirements.txt
django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pillow==12.0.0
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
dj-database-url==2.2.0
gunicorn==23.0.0
whitenoise==6.7.0
pytest==9.0.1
pytest-django==4.11.1
